Restructure the components files and create the courseform component.
