+++ src/components/course/courseForm.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function courseForm() {
+  return (
+    <div>courseForm</div>
+  );
+}
+
+export default courseForm;
