--- src/components/home.js
@@ -1,7 +1,0 @@
-function Home() {
-  return (
-    <div>this is the home page</div>
-  );
-}
-
-export default Home;
