+++ src/components/main_page/MainPage.css
@@ -1,0 +1,34 @@
+/* MainPage.css */
+.slider-container {
+  position: relative;
+}
+
+/* Style for previous button */
+.slick-prev {
+  position: absolute;
+  top: 50%;
+  left: -20px;
+  transform: translateY(-50%);
+  z-index: 1;
+}
+
+/* Style for next button */
+.slick-next {
+  position: absolute;
+  top: 50%;
+  right: -20px;
+  transform: translateY(-50%);
+  z-index: 1;
+}
+
+/* Circle background for each course */
+.circular-frame {
+  position: absolute;
+  background-color: #f0f4f8;
+  border-radius: 50%;
+  width: 60%;
+  height: 60%;
+  top: 5%;
+  left: 20%;
+  z-index: -1;
+}
